support building snap variants
We need to be able to build snaps with a suffix. These are identical
to our normal snap builds, but allow us to have a variation in
the store, like 'kubectl-foo'. This makes it possible to keep variants
on different versions without having a bunch of channels muddying up
the primary 'kubectl' snap.